backend/python-fastapi/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
passlib[bcrypt]
bcrypt==4.0.1
python-jose